rename client to frontend
