+++ frontend/src/App.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Provider } from "react-redux";
-import store from "store/index";
+import store from "store";
 import { createBrowserHistory } from "history";
 import { Router, Switch, Route } from "react-router-dom";
 import DefaultLayout from "layouts/DefaultLayout";
