+++ frontend/src/assets/projectImage.png
