+++ frontend/src/components/Polling/Card.jsx
