+++ frontend/src/components/Polling/SelectedPolls.jsx
