+++ frontend/src/layouts/AuthLayout.jsx
