+++ frontend/src/pages/JoinSession/index.jsx
