+++ frontend/src/pages/Polling/index.jsx
