+++ frontend/src/pages/SessionSelection/index.jsx
