+++ frontend/src/store/actions/polling.js
