+++ frontend/src/store/reducers/polling.js
